sidecar: extract shared brand, webview and autostart packages and add an external quit hook on Windows
