+++ sidecar/autostart_darwin.go
@@ -3,48 +3,22 @@
 package main
 
 import (
-	"fmt"
 	"os"
-	"path/filepath"
+
+	"github.com/jarvis/sidecar/internal/autostart"
 )
 
 // platformSetAutoStart writes (or removes) a per-user LaunchAgent plist so the
-// sidecar launches at login.
+// sidecar launches at login. Logic lives in internal/autostart so the
+// installer/uninstaller shares it.
 func platformSetAutoStart(enabled bool) error {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return err
-	}
-	dir := filepath.Join(home, "Library", "LaunchAgents")
-	plist := filepath.Join(dir, "com.jarvis.sidecar.plist")
-
-	if !enabled {
-		if err := os.Remove(plist); err != nil && !os.IsNotExist(err) {
+	exe := ""
+	if enabled {
+		var err error
+		exe, err = os.Executable()
+		if err != nil {
 			return err
 		}
-		return nil
 	}
-	exe, err := os.Executable()
-	if err != nil {
-		return err
-	}
-	if err := os.MkdirAll(dir, 0755); err != nil {
-		return err
-	}
-	content := fmt.Sprintf(`<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>Label</key>
-	<string>com.jarvis.sidecar</string>
-	<key>ProgramArguments</key>
-	<array>
-		<string>%s</string>
-	</array>
-	<key>RunAtLoad</key>
-	<true/>
-</dict>
-</plist>
-`, exe)
-	return os.WriteFile(plist, []byte(content), 0644)
+	return autostart.Set(exe, enabled)
 }
