+++ sidecar/autostart_windows.go
@@ -5,32 +5,20 @@ package main
 import (
 	"os"
 
-	"golang.org/x/sys/windows/registry"
+	"github.com/jarvis/sidecar/internal/autostart"
 )
 
-// autoStartValueName is the HKCU\...\Run value that points at this executable.
-const autoStartValueName = "JarvisSidecar"
-
-// platformSetAutoStart registers (or removes) the sidecar in the per-user
-// Windows "Run" key so it launches at login.
+// platformSetAutoStart registers (or removes) this executable in the per-user
+// Windows "Run" key so it launches at login. Logic lives in internal/autostart
+// so the installer/uninstaller shares it.
 func platformSetAutoStart(enabled bool) error {
-	k, err := registry.OpenKey(registry.CURRENT_USER,
-		`Software\Microsoft\Windows\CurrentVersion\Run`, registry.SET_VALUE)
-	if err != nil {
-		return err
-	}
-	defer k.Close()
-
-	if !enabled {
-		if err := k.DeleteValue(autoStartValueName); err != nil && err != registry.ErrNotExist {
+	exe := ""
+	if enabled {
+		var err error
+		exe, err = os.Executable()
+		if err != nil {
 			return err
 		}
-		return nil
 	}
-	exe, err := os.Executable()
-	if err != nil {
-		return err
-	}
-	// Quote the path so spaces in the install dir don't break the command.
-	return k.SetStringValue(autoStartValueName, `"`+exe+`"`)
+	return autostart.Set(exe, enabled)
 }
