+++ sidecar/internal/webview2/webview2_other.go
@@ -1,0 +1,10 @@
+//go:build !windows
+
+package webview2
+
+// Ensure is Windows-only; webview backends elsewhere (WebKitGTK, WKWebView)
+// ship with the OS or are package-manager dependencies.
+func Ensure() bool { return true }
+
+// Installed mirrors Ensure on non-Windows platforms.
+func Installed() bool { return true }
